feat(uaft): extract token watcher helper
